backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "einops>=0.8.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "openai>=1.102.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
